backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and Caching
redis==5.0.1
hiredis==2.3.2

# Celery for distributed tasks
celery==5.3.6
flower==2.0.1

# Git operations
GitPython==3.1.41
pygit2==1.14.0

# OpenAI
openai==1.10.0

# WebSocket
python-socketio==5.11.1
python-engineio==4.9.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Rate limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3

# Code analysis
radon==6.0.1
lizard==1.17.10
pylint==3.0.3

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
